modified the file 101-print_number.c again

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -9,7 +9,7 @@
  */
 void print_number(int n)
 {
-	long int i = n;
+	signed int i = n;
 
 	if (n < 0)
 	{
